requirements.txt
quarchpy
pyyaml
wmi